文档首页完成,管理员的用户管理完成

+++ src/main/java/org/ccclll777/alldocsbackend/controller/UserController.java
@@ -3,10 +3,15 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.ccclll777.alldocsbackend.entity.Category;
 import org.ccclll777.alldocsbackend.entity.User;
 import org.ccclll777.alldocsbackend.entity.dto.UserRegisterDTO;
+import org.ccclll777.alldocsbackend.entity.dto.UserRoleDTO;
 import org.ccclll777.alldocsbackend.entity.dto.UserUpdateDTO;
+import org.ccclll777.alldocsbackend.entity.vo.UserInfoVO;
 import org.ccclll777.alldocsbackend.enums.ErrorCode;
+import org.ccclll777.alldocsbackend.security.common.constants.SecurityConstants;
+import org.ccclll777.alldocsbackend.security.common.utils.JwtTokenUtils;
 import org.ccclll777.alldocsbackend.service.UserService;
 import org.ccclll777.alldocsbackend.utils.BaseApiResult;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,19 +29,11 @@ import java.util.List;
 public class UserController {
     @Autowired
     private  UserService userService;
-    @PostMapping("/sign-up")
+    @PostMapping("/register")
     @ApiOperation("用户注册")
     public BaseApiResult signUp(@RequestBody UserRegisterDTO userRegisterDTO) {
         userService.save(userRegisterDTO);
         return BaseApiResult.success("注册成功");
-    }
-    @GetMapping(value = "/getUsers")
-    @PreAuthorize("hasAnyRole('ROLE_MANAGER','ROLE_ADMIN')")
-    @ApiOperation("获取所有用户的信息（分页）")
-    public BaseApiResult getAllUser(@RequestParam(value = "pageNum", defaultValue = "0") int pageNum,
-                                    @RequestParam(value = "pageSize", defaultValue = "10") int pageSize) {
-        List<User> users = userService.selectUserList(pageNum, pageSize);
-        return BaseApiResult.success(users);
     }
     @PostMapping(value = "/insert")
     @PreAuthorize("hasAnyRole('ROLE_MANAGER','ROLE_ADMIN')")
@@ -68,7 +65,6 @@ public class UserController {
     @PreAuthorize("hasAnyRole('ROLE_MANAGER','ROLE_ADMIN','ROLE_USER')")
     @PostMapping(value = "/updateUser")
     public BaseApiResult updateUserInfo(@RequestBody UserUpdateDTO userUpdateDTO) {
-        log.info("更新用户入参==={}", userUpdateDTO.toString());
         User user = User.builder()
                 .id(userUpdateDTO.getId()).userName(userUpdateDTO.getUserName())
                 .nickName(userUpdateDTO.getNickName()).email(userUpdateDTO.getEmail())
@@ -77,12 +73,49 @@ public class UserController {
     }
     @ApiOperation(value = "根据id删除用户", notes = "根据id删除用户")
     @PreAuthorize("hasAnyRole('ROLE_MANAGER','ROLE_ADMIN')")
-    @DeleteMapping(value = "/deleteByUserId/{userId}")
-    public BaseApiResult deleteById( @PathVariable Integer userId) {
+    @DeleteMapping(value = "/delete/{userId}")
+    public BaseApiResult deleteById( @PathVariable Integer userId, @RequestHeader(SecurityConstants.TOKEN_HEADER) String token) {
+        String tokenValue = token.replace(SecurityConstants.TOKEN_PREFIX, "");
+        String tokenUserId =   JwtTokenUtils.getId(tokenValue);
+        if (Integer.parseInt(tokenUserId) == userId) {
+            return BaseApiResult.error(ErrorCode.USER_ERROR.getCode(), "无法删除自身");
+        }
         int row = userService.deleteUser(userId);
         if (row > 0) {
             return  BaseApiResult.success("删除成功");
         }
         return BaseApiResult.error(ErrorCode.DELETE_FAILE.getCode(),ErrorCode.DELETE_FAILE.getMessage());
+    }
+
+
+    @ApiOperation(value = "分页查询所有用户")
+    @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_MANAGER','ROLE_ADMIN')")
+    @GetMapping(value = "/all")
+    public BaseApiResult list(@RequestParam(value = "pageNum", defaultValue = "1") int pageNum,
+                              @RequestParam(value = "pageSize", defaultValue = "10") int pageSize) {
+        List<UserInfoVO> users = userService.selectUserList(pageNum-1, pageSize);
+        return BaseApiResult.success(users);
+    }
+
+    @ApiOperation(value = "查询用户数量")
+    @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_MANAGER','ROLE_ADMIN')")
+    @GetMapping(value = "/count")
+    public BaseApiResult count() {
+        int count = userService.getUserCount();
+        return BaseApiResult.success(count);
+    }
+
+    @ApiOperation(value = "更新用户权限")
+    @PreAuthorize("hasAnyRole('ROLE_MANAGER','ROLE_ADMIN')")
+    @PostMapping(value = "/updateUserRole")
+    public BaseApiResult updateUserRole(@RequestBody UserRoleDTO userRoleDTO,
+                                        @RequestHeader(SecurityConstants.TOKEN_HEADER) String token) {
+        System.out.println("12312312");
+        String tokenValue = token.replace(SecurityConstants.TOKEN_PREFIX, "");
+        String tokenUserId =   JwtTokenUtils.getId(tokenValue);
+        if (Integer.parseInt(tokenUserId) == userRoleDTO.getUserId()) {
+            return BaseApiResult.error(ErrorCode.USER_ERROR.getCode(), "无法修改自身权限");
+        }
+        return userService.updateUserRole(userRoleDTO.getUserId(),Integer.parseInt(userRoleDTO.getRoleId()));
     }
 }
